SEARCHKEY AND DIALOGS FINALISATION
