+++ backup.py
@@ -53,3 +53,28 @@ if __name__ == "__main__":
 
 
 
+#DIALOG PART 2
+from telegram.ext.conversationhandler import ConversationHandler
+
+from keyboards import keyboard_2_lvl_function
+
+from searchkey import searchkey_val
+
+
+def dialog_part_2(update, context):
+    print("dialog_part_2_activated")
+    user_answer_1 = update.message.text
+    context.user_data["dialog_dict1"] = {"key_part_1": user_answer_1}
+    print(context.user_data)
+    print("dialog_part_2_keyboard_activated")
+    keyboard_2_lvl_function (update, context)
+    return "user_temporary_dict_meal_subtype"
+
+def dialog_part_3(update, context):
+    print("dialog_part_3_activated")
+    user_answer_2 = update.message.text
+    context.user_data["dialog_dict2"] = {"key_part_2": user_answer_2}
+    print(context.user_data)
+    searchkey_val(update, context)
+    print(searchkey_val())
+    return ConversationHandler.END
